Rederive EC public key from mismatched private key.
If an EC public key (Q) does not match the private key (D), derive
Q again from D.

On Windows and Linux, if Q is entirely off the curve, an exception is
thrown. If Q is valid for the curve, but Q is not derived from the private
key and the generator, then it is recalculated.

On macOS, the public key Q is always re-derived from the private key.